Relocate undo components
