+++ src/main/java/seedu/address/logic/commands/modulelistcommands/UndoCommand.java
@@ -1,11 +1,13 @@
-package seedu.address.logic.commands;
-
-import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.Model;
+package seedu.address.logic.commands.modulelistcommands;
 
 import static java.util.Objects.requireNonNull;
 
-public class UndoCommand extends Command{
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.model.Model;
+
+public class UndoCommand extends Command {
     public static final String COMMAND_WORD = "undo";
     public static final String MESSAGE_USAGE = COMMAND_WORD
             + ": Undoes the last user command.\n"
